# api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "Category" : {
        "properties" : {
          "code" : {
            "example" : "GROCERIES",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "description" : "Only SPEND counts towards spent",
            "enum" : [ "SPEND", "INCOME", "TRANSFER" ],
            "type" : "string"
          },
          "name" : {
            "example" : "Groceries",
            "type" : "string"
          },
          "parentId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "code", "id", "kind", "name" ],
        "type" : "object"
      },
...
      "CategoryDetail" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "averageMinor" : {
            "description" : "Average over the trend months that have any data; absent when none has",
            "format" : "int64",
            "type" : "integer"
          },
          "category" : {
            "$ref" : "#/components/schemas/Category"
          },
          "currency" : {
            "example" : "RON",
            "type" : "string"
          },
          "merchants" : {
            "description" : "This month's variable spending by merchant (recurring charges excluded), largest first",
            "items" : {
              "$ref" : "#/components/schemas/MerchantAmount"
            },
            "type" : "array"
          },
          "month" : {
            "example" : "2026-03",
            "type" : "string"
          },
          "recurring" : {
            "description" : "This month's recurring payments, largest first",
            "items" : {
              "$ref" : "#/components/schemas/RecurringAmount"
            },
            "type" : "array"
          },
          "recurringMinor" : {
            "description" : "This month's recurring charges, total",
            "format" : "int64",
            "type" : "integer"
          },
          "trend" : {
            "description" : "The 12 months ending with `month`, oldest first",
            "items" : {
              "$ref" : "#/components/schemas/MonthAmount"
            },
            "type" : "array"
          }
        },
        "required" : [ "amountMinor", "category", "currency", "merchants", "month", "recurring", "recurringMinor", "trend" ],
        "type" : "object"
      },
...
      "MerchantAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "merchantId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "transactionCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "amountMinor", "merchantId", "name", "transactionCount" ],
        "type" : "object"
      },
...
      "MonthAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "hasData" : {
            "description" : "Whether any transaction exists in this month",
            "type" : "boolean"
          },
          "month" : {
            "example" : "2026-03",
            "type" : "string"
          }
        },
        "required" : [ "amountMinor", "hasData", "month" ],
        "type" : "object"
      },
...
      "RecurringAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "cadence" : {
            "enum" : [ "DAILY", "WEEKLY", "BIWEEKLY", "MONTHLY", "QUARTERLY", "YEARLY" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "state" : {
            "description" : "CONFIRMED or ENDED",
            "type" : "string"
          },
          "subscriptionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "transactionCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "amountMinor", "cadence", "name", "state", "subscriptionId", "transactionCount" ],
        "type" : "object"
      },
...
      "TransactionList" : {
        "properties" : {
          "count" : {
            "format" : "int32",
            "type" : "integer"
          },
          "currency" : {
            "example" : "RON",
            "type" : "string"
          },
          "items" : {
            "description" : "Newest first",
            "items" : {
              "$ref" : "#/components/schemas/TransactionView"
            },
            "type" : "array"
          },
          "month" : {
            "description" : "The period's last month",
            "example" : "2026-03",
            "type" : "string"
          },
          "months" : {
            "description" : "1, or 3 for the 3 months ending with `month`",
            "format" : "int32",
            "type" : "integer"
          },
          "scope" : {
            "enum" : [ "SPEND", "INCOME", "ALL" ],
            "type" : "string"
          },
          "totalMinor" : {
            "description" : "SPEND: money spent (positive); INCOME: money in; ALL: net (in − out)",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "count", "currency", "items", "month", "months", "scope", "totalMinor" ],
        "type" : "object"
      },
      "TransactionView" : {
        "properties" : {
          "accountId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "amountMinor" : {
            "description" : "Signed minor units; negative = money out",
            "format" : "int64",
            "type" : "integer"
          },
          "bookingDate" : {
            "format" : "date",
            "type" : "string"
          },
          "categoryCode" : {
            "type" : "string"
          },
          "categoryConfidence" : {
            "type" : "number"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categorySource" : {
            "description" : "USER, RULE, LEARNED, MCC, KEYWORD, SYSTEM; absent when uncategorized",
            "type" : "string"
          },
          "currency" : {
            "example" : "RON",
            "type" : "string"
          },
          "description" : {
            "description" : "Bank text with IBANs masked",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "merchantKey" : {
            "type" : "string"
          },
          "merchantName" : {
            "type" : "string"
          },
          "status" : {
            "description" : "PENDING: not posted yet (e.g. a card reservation); may still change",
            "type" : "string"
          },
          "subscriptionId" : {
            "description" : "The recurring payment (or proposal) this is a charge of",
            "format" : "int64",
            "type" : "integer"
          },
          "subscriptionName" : {
            "type" : "string"
          },
          "subscriptionState" : {
            "description" : "PROPOSED, CONFIRMED or ENDED",
            "type" : "string"
          },
          "transferAccountName" : {
            "description" : "The other own account of a transfer",
            "type" : "string"
          },
          "transferState" : {
            "description" : "PAIRED: a transfer between own accounts, both sides seen; PROVISIONAL: only this side, recognised by the other account's IBAN; absent: not an own-account transfer",
            "type" : "string"
          }
        },
        "required" : [ "accountId", "amountMinor", "bookingDate", "currency", "description", "id", "merchantKey", "merchantName", "status" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "Outflow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/insights/categories/{id}" : {
      "get" : {
        "operationId" : "category",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "YYYY-MM",
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "ISO currency; absent = the main currency of the accounts (most transactions)",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Account ids to include (accounts filter); absent = all accounts",
          "in" : "query",
          "name" : "accounts",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insights" ]
      }
    },
...
    "/api/transactions" : {
      "get" : {
        "operationId" : "list_1",
        "parameters" : [ {
          "description" : "YYYY-MM",
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "ISO currency; absent = the main currency of the accounts (most transactions)",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "scope",
          "required" : false,
          "schema" : {
            "default" : "ALL",
            "enum" : [ "SPEND", "INCOME", "ALL" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "category",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "uncategorized",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "merchant",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "Merchant or bank text contains it, or the amount equals it",
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "1, or 3 for the 'last 3 months' view ending with the month",
          "in" : "query",
          "name" : "months",
          "required" : false,
          "schema" : {
            "default" : 1,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "description" : "Account ids to include (accounts filter); absent = all accounts",
          "in" : "query",
          "name" : "accounts",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          }
        }, {
          "description" : "true: only charges of confirmed or ended recurring payments; false: only the rest",
          "in" : "query",
          "name" : "recurring",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "description" : "Only the charges of this recurring payment",
          "in" : "query",
          "name" : "subscription",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionList"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transactions" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}